Refactor: Changed file structure to colocate components and data with the feature route itself
