+++ src/app/(home)/_components/Faqs.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import React, { useState } from 'react';
-import faqInfo from '@/data/faqs';
-import { H1, Section } from '../common/Typography';
+import faqInfo from '@/app/(home)/_data/faqs';
+import { H1, Section } from '../../../components/common/Typography';
 
 const Faqs = () => {
   const [openAccordion, setOpenAccordion] = useState<number | null>(null);
@@ -12,7 +12,7 @@ const Faqs = () => {
   };
 
   return (
-    <Section id="faqs" className='mb-10'>
+    <Section id="faqs" className="mb-10">
       <H1 className="mb-8">FAQs</H1>
       <div className="flex flex-col space-y-4">
         {faqInfo.map((faq, index) => (
@@ -23,13 +23,13 @@ const Faqs = () => {
             } rounded-lg border border-gray-300 bg-white`}
           >
             <div
-              className="collapse-title cursor-pointer text-xl lg:text-2xl font-medium text-black"
+              className="collapse-title cursor-pointer text-xl font-medium text-black lg:text-2xl"
               onClick={() => handleToggle(index)}
             >
               {faq.question}
             </div>
             <div className="collapse-content">
-              <p className="text-black text-lg lg:text-xl">{faq.answer}</p>
+              <p className="text-lg text-black lg:text-xl">{faq.answer}</p>
             </div>
           </div>
         ))}
