+++ src/app/(home)/_components/sponsors/SponsorCard.tsx
