+++ src/app/(home)/_data/faqs.ts
