+++ src/app/organisers/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import React from "react";
+import React from 'react';
 import { motion } from 'motion/react';
-import organisers from "@/data/organisers";
+import organisers from '@/app/organisers/_data/organisers';
 
 interface TeamCardProps {
   img: string | null;
@@ -14,7 +14,7 @@ interface TeamCardProps {
 function TeamCard({ img, name, title, club }: TeamCardProps) {
   return (
     <motion.div
-      className="group relative mx-auto w-48 rounded-lg bg-black shadow-md overflow-hidden"
+      className="group relative mx-auto w-48 overflow-hidden rounded-lg bg-black shadow-md"
       variants={{
         hidden: { opacity: 0, y: 50 },
         visible: { opacity: 1, y: 0, transition: { duration: 0.5 } },
@@ -23,22 +23,22 @@ function TeamCard({ img, name, title, club }: TeamCardProps) {
       {/* Image */}
       <div className="h-48 w-full overflow-hidden">
         <img
-          src={img || "https://via.placeholder.com/96"}
+          src={img || 'https://via.placeholder.com/96'}
           alt={name}
           className="h-full w-full object-cover object-top transition-opacity duration-300 group-hover:opacity-50"
         />
       </div>
 
       {/* Info */}
-      <div className="absolute inset-0 flex flex-col items-center justify-center bg-gray-0 bg-opacity-0 text-center opacity-0 transition-opacity duration-300 group-hover:opacity-100">
+      <div className="bg-gray-0 absolute inset-0 flex flex-col items-center justify-center bg-opacity-0 text-center opacity-0 transition-opacity duration-300 group-hover:opacity-100">
         <h5 className="text-xl font-semibold text-white">{name}</h5>
-        <p className="mb-2 text-sm font-semibold text-white text-center break-words w-full max-w-[180px]">
+        <p className="mb-2 w-full max-w-[180px] break-words text-center text-sm font-semibold text-white">
           {title}
         </p>
         <img
           src={club}
           alt={`${name}'s club`}
-          className="h-8 w-20 mt-2" // Adjust size as needed
+          className="mt-2 h-8 w-20" // Adjust size as needed
         />
       </div>
     </motion.div>
@@ -57,7 +57,7 @@ function TeamSection() {
   };
 
   return (
-    <section className="min-h-screen py-8 px-8 lg:py-28">
+    <section className="min-h-screen px-8 py-8 lg:py-28">
       <div className="container mx-auto max-w-screen-sm">
         <div className="mb-16 text-center lg:mb-28">
           <h1 className="my-2 text-2xl font-bold text-white lg:text-4xl">
@@ -65,7 +65,7 @@ function TeamSection() {
           </h1>
         </div>
         <motion.div
-          className="grid grid-cols-1 gap-6 mt-6 justify-center sm:grid-cols-2 lg:grid-cols-3"
+          className="mt-6 grid grid-cols-1 justify-center gap-6 sm:grid-cols-2 lg:grid-cols-3"
           variants={containerVariants}
           initial="hidden"
           animate="visible"
