+++ src/app/past-winners/_data/past-winners/pastWinners2022.ts
