+++ src/app/past-winners/_data/pastWinners.ts
