+++ src/app/registration/_data/registrationFaqs.ts
