--- src/components/Sponsors.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const Sponsors = () => {
-  return <div>Sponsors</div>;
-};
