--- src/types/type.ts
@@ -1,1 +1,0 @@
-//placeholder
